separated out Experimental & Theoretical scattering
